* master:
  scripts/test-persistent-reservation: Enable APTPL
  fcst: Fix a recently introduced compiler warning
  Increase version to -pre2. Pre1 was staying for too long.
  Docs and web updates
  cmd's bufflen and data_len overflow cleanups
  ib_srpt: Convert two functions two single-exit style
  Avoid that enabling a target too quickly triggers a crash
  ib_srpt README: Add more performance notes
  Fix possible BUG_ON() on multiple forced commands release
  Fix scst_local crash on 0 response data len w/o data buffer allocated
  ib_srpt README: Add more performance notes
  qla2x00t: Display a clear error message if the kernel is too old
  scst: Document CONFIG_SCST_DEBUG_SYSFS_EAGAIN
  srpt/README: Recommend to use SLAB instead of SLUB
  srpt/README: Recommend to disable CONFIG_SCHED_DEBUG and CONFIG_SCHEDSTATS
  scst_user: Change a goto statement into an explicit loop
  The scst_unregister_target() call frees *tgt and hence also tgt->tgt_list.next. Avoid that unregistering all targets triggers a use-after-free of tgt->tgt_list.next. Found this through source reading.
  The comment above scst_act.acg_list_entry was correct until the sysfs interface was introduced. Update it with information about how acg_list_entry is used in the sysfs interface.
  ib_srpt: Update README

+++ scst/include/scst.h
@@ -2041,7 +2041,7 @@ struct scst_cmd {
 	 * VERIFY, which transfer different amount of data (if any), than
 	 * processed.
 	 */
-	int data_len;
+	int64_t data_len;
 
 	/* Completion routine */
 	void (*scst_cmd_done) (struct scst_cmd *cmd, int next_state,
@@ -2633,7 +2633,7 @@ struct scst_acg {
 	/* List of attached acn's, protected by scst_mutex */
 	struct list_head acn_list;
 
-	/* List entry in acg_lists */
+	/* List entry in acg_lists (procfs) or tgt_acg_list (sysfs) */
 	struct list_head acg_list_entry;
 
 	/* Name of this acg */
@@ -3161,7 +3161,7 @@ static inline int scst_cmd_get_bufflen(struct scst_cmd *cmd)
  * Returns cmd's data_len. See the corresponding field's description in
  * struct scst_cmd above.
  */
-static inline int scst_cmd_get_data_len(struct scst_cmd *cmd)
+static inline int64_t scst_cmd_get_data_len(struct scst_cmd *cmd)
 {
 	return cmd->data_len;
 }
